remove lua as a dependency
Thanks to a tip from Ally, apparently, Ouch Browser can compile and run just
fine without the Lua interpreter or the development libraries installed on
the system. This is because mlua is compiled in "vendored" mode, meaning
that Ouch Browser will always be compiled with the interpreter and
development libraries built-in.